Добавил фрагмент и функцию для просмотра полной информации о фильме.

+++ app/src/main/java/com/example/kplist/data/models/apiMovieModel/World.kt
@@ -3,5 +3,5 @@ package com.example.kplist.data.models.apiMovieModel
 data class World(
     val _id: String,
     val currency: String,
-    val value: Int
+    val value: Long
 )
